requirements.txt
Flask
gunicorn
kafka-python
psycopg2-binary
PyYAML
prometheus_client
pytest